feat(Mobile): Convert to mobile app

+++ src/utils/api.ts
@@ -1,9 +1,14 @@
 import axios from "axios";
+import { Capacitor, CapacitorHttp as Http } from "@capacitor/core";
 
-const baseURL: string = import.meta.env.VG_API_BASE_URL;
+const isMobile = Capacitor.isNativePlatform();
+
+export const API_URL: string = isMobile
+	? import.meta.env.VG_API_BASE_URL_MOBILE || import.meta.env.VG_API_BASE_URL
+	: import.meta.env.VG_API_BASE_URL;
 
 const api = axios.create({
-	baseURL,
+	baseURL: API_URL,
 	withCredentials: true,
 	headers: {
 		"Content-Type": "application/json",
@@ -11,7 +16,7 @@ const api = axios.create({
 });
 
 const apiFile = axios.create({
-	baseURL,
+	baseURL: API_URL,
 	headers: {
 		"Content-Type": "multipart/form-data",
 	},
@@ -35,33 +40,64 @@ api.interceptors.response.use(
 );
 
 export const getData = async (url: string) => {
+	if (isMobile) {
+		const { data } = await Http.get({
+			url: `${API_URL}${url}`,
+			headers: { "Content-Type": "application/json" },
+		});
+		return data;
+	}
 	const { data } = await api.get(url);
 	return data;
 };
 
-// export const getData = async (url: string, controller?: AbortController) => {
-// 	const response = await fetch(`http://localhost:3000/v1${url}`);
-// 	const resp = await response.json();
-// 	console.log("dataaa", resp);
-// 	return resp;
-// };
-
 export const postData = async (url: string, post: object) => {
+	if (isMobile) {
+		const { data } = await Http.post({
+			url: `${API_URL}${url}`,
+			headers: { "Content-Type": "application/json" },
+			data: post,
+		});
+		return data;
+	}
 	const { data } = await api.post(url, post);
 	return data;
 };
 
 export const patchData = async (url: string, post: object) => {
+	if (isMobile) {
+		const { data } = await Http.patch({
+			url: `${API_URL}${url}`,
+			headers: { "Content-Type": "application/json" },
+			data: post,
+		});
+		return data;
+	}
 	const { data } = await api.patch(url, post);
 	return data;
 };
 
 export const deleteData = async (url: string) => {
+	if (isMobile) {
+		const { data } = await Http.delete({
+			url: `${API_URL}${url}`,
+			headers: { "Content-Type": "application/json" },
+		});
+		return data;
+	}
 	const { data } = await api.delete(url);
 	return data;
 };
 
-export const uploadFile = async (url: string, post: object) => {
+export const uploadFile = async (url: string, post: FormData) => {
+	if (isMobile) {
+		const { data } = await Http.post({
+			url: `${API_URL}${url}`,
+			headers: { "Content-Type": "multipart/form-data" },
+			data: post,
+		});
+		return data;
+	}
 	const { data } = await apiFile.post(url, post);
 	return data;
 };
